style(php-cs-fixer): Fix formatting

+++ packages/sanf/api/src/Modules/Branch/DetailBranchTransformer.php
@@ -1,25 +1,22 @@
 <?php
 
-
 namespace Sanf\Api\Modules\Branch;
-
 
 use League\Fractal\TransformerAbstract;
 
 class DetailBranchTransformer extends TransformerAbstract
 {
-
     public function transform($dto)
     {
         return [
-            'id' => (int)$dto->id,
-            'name' => (string)$dto->name,
-            'address' => (string)$dto->address,
-            'msisdn' => (string)$dto->msisdn,
-            'msisdn_alternative' => (string)$dto->msisdn_alternative,
-            'email' => (string)$dto->email,
-            'latitude' => (double)$dto->latitude,
-            'longitude' => (double)$dto->longitude,
+            'id' => (int) $dto->id,
+            'name' => (string) $dto->name,
+            'address' => (string) $dto->address,
+            'msisdn' => (string) $dto->msisdn,
+            'msisdn_alternative' => (string) $dto->msisdn_alternative,
+            'email' => (string) $dto->email,
+            'latitude' => (float) $dto->latitude,
+            'longitude' => (float) $dto->longitude,
         ];
     }
 }
